Make Model "updateSize" methods not require a user
Access controls never have an impact on the "size" of Collections, Users,
Folders, or Items. Sizes for Collections and Users represent the total
size of all content recursively within them, regardless of whether a given
user has access to any of it.

As a practical matter, the system consistancy check which typically calls
"updateSize" may only be run by a site admin user, who will have access
to everything.