Remove user access checks from _fixBaseParents in consistancy check
A user may have access to a sub-folder that is within a parent folder or
collection that the user does not have access to (this is how "unlisted"
folders can be implemented). Regardless, these sub-folders should always
have an accurate "baseParentId" and "baseParentType", even if a user does
not have rights to actually traverse to this base.

As a practical matter, the system consistancy check may only be run by a
site admin user, who will have access to everything.