Must pipe pug templates through babel
Now that we are using ES2015 features like string templates,
we have to post-process the pug output through babel in order
for it to work on ES5-only browsers (such as phantom).

Notable changes: Had to change octal notation to the ES2015
standard, and had to remove the variable "public" from templates
due to it being a reserved word.