Support static-only plugin dependencies
This allows plugins to build web content that depends on web
content from other plugins, but have no runtime dependence on
the other plugin.

The motivating example for this is that slicer_cli_web and
item_tasks share web client code, but do not have any runtime
requirements, and probably won't be deployed together.