fontello.zip sometimes fails to download
When running the grunt tasks (such as via `npm install`), fontello.zip is downloaded from data.kitware.com.  Frequently, this fails, and an error of ` Corrupted zip : can't find end of central directory` is emitted when the file is unzipped.  Rerunning the task will often succeed (it has the same chance of failure as the initial run).

Anecdotally, this seems to happen more often when running in a VM or docker environment.

This appears to be caused by nodejs's `request` module emitting a finished event before the file is fully flushed to disk via a stream made with `fs.createWriteStream`.  This happens in the `grunt-curl` task (which, misleadingly, doesn't use `curl`), and also happens with the same frequency if the `request` module is used alone.

As a solution, we can shell out to the actual `curl` command.  We already require curl as a dependency, according to our docs.