Adjust the host in the swagger document.
When serving the api describe endpoint, try to use the referrer to set the host and basePath in the swagger document.  This is well-formed from most web browsers, and will properly work inside and outside of a proxy.  For requests that don't send a Referer [sic] header, use the existing method.  Without this, proxies will only have the right value in Swagger if the tools.proxy.base value is set, if the api root is not placed on a different path by the proxy, and if the proxy is accessed as described by the tools.proxy.base value (which might be different inside and outside of a vpn, for instance).