Fix an issue getting the access level of an item.
Also, allow an access level to be passed to the item list widget so that access levels don't have to be retrieved.  This is used by the Hierarchy Widget.

If the access level of an item is unknown, the parent of the item (a folder) is retrieved.  If the access level was requested again while the parent was still being fetched, undefined would be returned instead of waiting for the fetch to complete.  Now, a new fetch is started.  This is wasteful, for further improvements could be to wait for the first fetch to complete or to share a single parent item between all of the items in a collection.  The fetch is avoided completely for the Hierarchy Widget by assigning an access level when the item list is created.