Change a style of the user text links.
On the header bar, when you are logged in, the entire height around the user name acts as the click target for the link.  When you are logged out, the Register and Log In links' click targets are just the height of the text.  This makes the logged out click targets the full height.