Remove confirm event handler when modal dialog is hidden
The 'click' event handler in confirm() held a reference to
params.confirmCallback, which in turn could hold references to objects
in the view that set up the callback. In certain cases this led to leaks
of models, widgets, and DOM elements referenced by the view.

One example where this was a problem is navigating to the group list,
selecting a group, selecting the menu item to delete the group,
canceling, and navigating back to, say, the front page. At this point,
it's expected that resources related to the group view and confirm
dialog have been destroyed.