Destroy jqplot in QuotaPoliciesWidget
It's necessary to explicitly release the resources of a jqPlot plot.
This avoids leaking DOM trees and potentially other objects.

This commit also ensures that QuotaPoliciesWidget is destroyed
immediately after the modal dialog is hidden. Otherwise, if a user views
the quota policies dialog more than once, multiple widgets are created
and none are destroyed until the parent view is destroyed. Attempting to
release the resources of the jqPlot plots at that point can result in an
error.