Support single POST chunk uploads.
If the parameters on a POST /file endpoint are included in the URL and the body is not processed by cherrypy and exists, then use the body as the first chunk of the upload.  Alternately, if there is a `chunk` parameter, use that as the first chunk of the upload.

If multiple chunks are needed, the first can be sent via POST /file, and the rest must be sent via POS /file/chunk.

If the file is completely uploaded, POST /file will return a file resource.  Otherwise, it will return an upload resource.

Skip saving an upload model when the file is uploaded as a single chunk via the POST /file endpoint.  If only the first chunk of multiple chunks is sent via the POST /file endpoint, then the upload model will be saved at the end of saving the first chunk.

Once this is merged to Girder master and there exists a new version, the girder client can be modified to use a single POST for small uploads (or always start with a POST for the first part of an upload).

This resolves issue #2088.