Trigger save event when updating model ACL
This fixes a (maybe breaking?) change in behavior as a result
of how we update a document's ACL; until recently, we called
save() during this process if save was set to True, which
caused the save events for the model to trigger.

Specifically, this broke some expectations of COVALIC where we
were expecting the ACL updates to trigger save events.