Unify importing tasks from JSON and Slicer CLI
Previously it was possible to:
- Configure an item as a task from a Slicer CLI container.
- Add task items to a folder from a JSON container.

This change adds functionality to do both actions from either type of
container. The new configure tasks dialog allows the user to choose
between a JSON container and a Slicer CLI container by choosing a tab.

On the server, this change adds new endpoints to support the new
functionality. The endpoints extend the existing 'item' and 'folder'
routes so that the model type expected by the id URL parameter is clear.
The old endpoints are deprecated, but not removed.

Note that minor updates may necessary once support for Slicer CLI
containers describing more than one item is finalized.