Raise exceptions more correctly in GirderClient.authenticate
This commit causes GirderClient.authenticate to raise an
AuthenticationError only in the case of 401 and 403 status codes,
while raising an HttpError for all others, regardless of whether or
not 'authToken' is in the response.

An instance of this behavior causing problems is a scenario in which a
response returns a 500. In this case the correct exception would never
be thrown. If it happened to have had a JSON response it would raise
an AuthenticationError (technically correct, though the error message
would be indistinguishable from a 401/403). If it didn't have a JSON
response, an AttributeError would be raised to the caller giving no
indication as to what happened on the server side.

This commit also removes a comment about following redirects since
that is enabled by default (see the requests.request flag
allow_redirects).