Deprecate girder_client.HttpError
This commit deprecates HttpError in favor of requests.HTTPError which
Girder will be fully transitioning to in version 3.0.

Additionally, the constructor for raising an HttpError now accepts a
response object to allow requests.HTTPError to set request/response on
the exception. Further use of HttpError should include this to allow
callers to use exc.response and exc.request.

See https://github.com/girder/girder/pull/2223#issuecomment-318636072
for rationale.