Add support for image parameter types in item_tasks
Before this change slicer parameters of type "image" were added
to the item_tasks spec as type "file".  The image type is essentially
just an alias of the file type (which is really an item...).  It
is useful to distinguish between them in HistomicsTK because images
can be opened by the large_image viewer.

This would probably be a good candidate for a parameter type that
is added as an extension to item_tasks, but we need to do some
refactoring to support dynamic extensions before that is possible.