Address caching getCurrentUser
Caching getCurrentUser with dogpile is incompatible due to the fact
that the dogpile caching layer can be enabled or disabled.
getCurrentUser relies on the ability of Girder to explicitly set the
current user with setCurrentUser (and we have code paths which rely on
this). This means that _cacheAuthUser wasn't acting so much as a cache
mechanism as it was a means for providing getCurrentUser with the true
value (from setCurrentUser).