Partially reverts "97bdf57886d0a847c"
This fixes an issue where the event daemon was never actually being
terminated. Coverage shows while self.terminate gets set to True, the
log message printing the stopped message never actually gets run. This
is due to a bug where we indefinitely block for another item off the
queue in the middle of our 'while not terminate' clause.