Fix the terms client test.
When we use `window.location.assign`, make sure to wait for something that we expect to change.  Just doing `girderTest.waitForLoad()` is insufficient, since window.assign doesn't cause the change to start until the next time slice, and `waitForLoad` will return immediately that things are loaded (probably a `waits(1)` would have worked instead of the `waitsFor` that was added, but that seems more brittle).

Once we've properly waited for the `window.location.assign`, fix waiting for creating an item.  A comment was added that waiting caused a test to fail, which was a good indication that we *weren't* waiting for the correct thing.

Also, fix throwing an error in the test utilities to use the error constructor.