Harden routing tests.
Switch the order of some of the routing test tests.  Before, we waited for rest requests to be done then for the routing test-specific test.  This now waits for the routing-specific test first.  If there is a delay between routing and any of the loading actions, no rest requests might be active, the routing condition is then detected, and then the loading actions start.  Future tests may fail because they expect the loading actions to be complete and initiate another action while.

In testing this repeatedly, I saw 4 failures in 250 tests before this change and 0 after.  I don't know if this will resolve all failure cases, but it does harden the test and should have no ill effects.