Various fixes for testing plugins in pytest
The main purpose of this commit is to put band-aids over several
problems that all result from a single bad assumption, which
is treating the girder server as a single global python object.
In the old testing mode, we only ever had one cherrypy server per
test process, so we didn't run into these issues where we treated
the python global namespace as equivalent to the context of the
cherrypy server.

In our new pytest framework, we need to spin the entire server
context up and down repeatedly in the same python global context,
meaning we either need the server fixture cleanup process to
tear everything down to a good state, or to make the server startup
process idempotent, or (ideally) both. This commit fixes the issues I ran
into, and does a little bit of both.

I'd like to discuss a long-term / robust solution to this issue
at some point soon.