Allow the pythonpath to be adjusted for the client test module.
We added the ability to specify the client test module in PR #1254.  In
that PR, we started passing whatever the PYTHONPATH environment variable
was to the test.  Although this allows *one* test to override the path,
it fails when there are two such tests in different directories.  This
allows an additional parameter (TEST_PYTHONPATH) to be added to aid in
overriding the PYTHONPATH.  In this way, two tests that require
different paths can be run in the same build.