Add a click event on the App.
Adding hrefs to elements and eliminating explicit routing has a few
repercussions, which this change should fix.

First, calling `$(elem).click()` from javascript will only work if there
is a jQuery event (including Backbone events) on the element or a parent
with the appropriate selector.  Removing those events disables the
ability to use the jQuery click event to trigger a click.  As a
fallback, one can use `$(elem)[0].click()`, but this would require
changing a large number of tests.

Rather, this creates a new Backbone event on the app which handles
clicks on `a` elements which have `href` attributes that are anchor
fragments (start with `#`).  This will only operate if no lower event
has not prevented the default (to avoid navigating twice).  This means
that the user clicking on a href element *and* calling `$(elem).click()`
*will* do the same thing.

Further, we had been mildly abusing the `li` elements on the admin
console.  Instead of having the `a` element be the area the user should
click, it was just an interior div and the wrapping `li` was the real
link.  This changes that -- the `a` element contains the desired padding
and fills the `li` element completely (thus making the `a` element be
the target click area).  By shifting where the classes are defined, no
tests need to change.

As a result of this commit, (a) all tests are unchanged from master, (b)
we have less custom routing code, (c) we have hrefs on a number of links
that didn't have them.  Further, testing other `a` hrefs should be more
convenient, since the jQuery `click` handler will work on them now
(without monkeying with jQuery's events).