Provide a sensible default for events.daemon
Some of our own tests assumed that events.daemon us setup outside
the context of the Girder server, so it's quite possible that
downstream tests also make this assumption. To be less disruptive,
we use a default value for the daemon.

I'm not sure if it's more sensible to use the foreground handler
as the default, or the threaded mode.