Report if we can't unmount fuse.
If any Girder file is still open when unmount is called, unmount will
fail.  Instead of quietly trying to end, report the error code so the
unmounting caller can then choose what to do.

Release a file handle in one of the tests.  Prior to this change, there
were some wait loops in the tests to try to clean up that still
sometimes could fail due to the file being open and the mount not
closing.