Refactor session logic in girder_client cli
The original code that injected a session into all requests correctly
created a request method on a session, but the session itself was closed
by the context manager before the method was called.  This just happened
to work for the ssl verification flag, but would not work when trying to
inject additional adapters.

To ensure the session is actually used for all rest requests, this
changes all calls to _requestFunc to go through the sendRestRequest
method instead.  This allows the girder_client cli to override the
sendRestRequest and ensure an open session is used for the duration of
the request.