Remove temporary form after submitting
This commit fixes a problem in the web client where a text input appears
at the bottom of the page after checking items and selecting "Download
checked resources".

To avoid long query strings in GET requests--in case a large number of
files are checked--this download mechanism creates and submits a
temporary form. For compatibility with Firefox, the form must be in the
DOM for submission to work.

This commit removes the temporary form from the DOM after it's
submitted.