Harden client tests by changing waitForLoad.
We have seen intermittent failures of the terms plugin client test.  In
tracking this down, when the test logs in as a user which viewing an
item page, sometimes the rest calls all happen within the first few
javascript time slices, and then the waitForLoad function properly waits
until all rest calls are complete.  The list of files for the item is
not called until the render method for the item is called.  This doesn't
have to be in the same time slice, and, if all other actions have
finished, waitForLoad returns.  When the list of files is fetched, the
page renders the item, which changes what is visible and causes
subsequent tests to fail.

This changes how waitForLoad waits.  Before it waited until all
transitions where finished, no more rest requests were outstanding, and
all loading blocks were replaced by their results.  This now adds
waiting a few time slices where all of this is true.  If any of the
previous conditions fail, the counter is reset.  This means that, in
general, even if a render is a time slice or two behind load completion
and triggers a new rest request, waitForLoad will now correctly wait for
total completion.

As a local test, before, running `ctest --repeat-until-fail 100 -R
web_client_terms.terms` never succeeded, and typically fails within the
first 10 tests (though, rarely, it gets through 20 tests).  After this
change, the same command has succeeded twice.